requirements.txt
Flask==0.12.2
Flask-Cors==3.0.3
Flask-PyMongo==0.5.1
gunicorn==19.7.1
`numpy==1.13.3
numpydoc==0.7.0
scipy==0.19.1